+++ NEWS.md
@@ -1,5 +1,8 @@
-# enaho 0.2.1
-* variablesxperiodo
+# enaho 0.2.5
+* Added `media.enaho()`.
+* Added `proporcion.enaho()`.
+* Added `variablesxperiodo()`.
+* Added new vignettes.
 
 # enaho 0.2.0